Commit changes in the order of the classes
