--- AboutPage.xaml.cs
@@ -1,13 +1,0 @@
-﻿
-using Xamarin.Forms;
-
-namespace MoviesListProject.Views
-{
-    public partial class AboutPage : ContentPage
-    {
-        public AboutPage()
-        {
-            InitializeComponent();
-        }
-    }
-}
