--- AboutViewModel.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Windows.Input;
-using Xamarin.Forms;
-
-namespace MoviesListProject.ViewModels
-{
-    public class AboutViewModel : BaseViewModel
-    {
-        public AboutViewModel()
-        {
-            Title = "About";
-
-            OpenWebCommand = new Command(() => Device.OpenUri(new Uri("https://xamarin.com/platform")));
-        }
-
-        /// <summary>
-        /// Command to open browser to xamarin.com
-        /// </summary>
-        public ICommand OpenWebCommand { get; }
-    }
-}
