--- MovieDetailsPage.xaml
@@ -1,36 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="MoviesListProject.Views.MovieDetailsPage">
-    <ContentPage.Content>
-        <StackLayout>
-            <StackLayout Orientation="Horizontal">
-                <Label Text="Title: " />
-                <Label Text="{Binding Title}" />
-            </StackLayout>
-            <StackLayout Orientation="Horizontal">
-                <Label Text="Original Title: " />
-                <Label Text="{Binding movie.OriginalTitle}" />
-            </StackLayout>
-            <StackLayout Orientation="Horizontal">
-                <Label Text="Release Date: " />
-                <Label Text="{Binding movie.ReleaseDate}" />
-            </StackLayout>
-            <StackLayout Orientation="Vertical">
-                <Label Text="Overview: " />
-                <Label Text="{Binding movie.Overview}" />
-            </StackLayout>
-            <StackLayout Orientation="Horizontal">
-                <Label Text="Genre: " />
-                <Label Text="{Binding MoviePrincipalGenre}" />
-            </StackLayout>
-            <StackLayout Orientation="Horizontal">
-                <Label Text="Poster: " />
-            <Image x:Name="imgPoster"
-                   HorizontalOptions="Start"
-                   Source="{Binding movie.MoviePosterPath}"
-                   WidthRequest="150"></Image>
-            </StackLayout>
-        </StackLayout>
-    </ContentPage.Content>
-</ContentPage>
