--- MoviesListPage.xaml
@@ -1,50 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="MoviesListProject.Views.MoviesListPage"
-             Title="{Binding Title}"
-             x:Name="MoviesListPage">
-    <ContentPage.ToolbarItems>
-    </ContentPage.ToolbarItems>
-    <ContentPage.Content>
-        <StackLayout>
-            <ListView x:Name="MoviesListView"
-                      ItemsSource="{Binding Movies}"
-                      VerticalOptions="FillAndExpand"
-                      HasUnevenRows="true"
-                      IsPullToRefreshEnabled="true"
-                      IsRefreshing="{Binding IsBusy, Mode=OneWay}"
-                      CachingStrategy="RecycleElement"
-                      SelectedItem="{Binding SelectedItem}">
-                <ListView.ItemTemplate>
-                    <DataTemplate>
-                        <ViewCell>
-                            <StackLayout Padding="10">
-                                <Label Text="{Binding Title}"
-                                       LineBreakMode="NoWrap"
-                                       Style="{DynamicResource ListItemTextStyle}"
-                                       FontSize="16" />
-                                <StackLayout Orientation="Horizontal">
-                                    <Label Text="Release Date: "
-                                           LineBreakMode="NoWrap"
-                                           Style="{DynamicResource ListItemDetailTextStyle}"
-                                           FontSize="13" />
-                                    <Label Text="{Binding ReleaseDate}"
-                                           LineBreakMode="NoWrap"
-                                           Style="{DynamicResource ListItemDetailTextStyle}"
-                                           FontSize="13" />
-                                </StackLayout>
-                            </StackLayout>
-                        </ViewCell>
-                    </DataTemplate>
-                </ListView.ItemTemplate>
-            </ListView>
-            <StackLayout Orientation="Horizontal">
-                <Button Command="{Binding BackCommand}"
-                        Text="Back Page" />
-                <Button Command="{Binding NextCommand}"
-                        Text="Next Page" />
-            </StackLayout>
-        </StackLayout>
-    </ContentPage.Content>
-</ContentPage>
