+++ MoviesListProject/MoviesListProject/ViewModels/MoviesListViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using MoviesListProject.Helpers;
+using MoviesListProject.Views;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
@@ -13,32 +14,80 @@ namespace MoviesListProject.ViewModels
 {
     public class MoviesListViewModel : BaseViewModel
     {
+        public ICommandEx NextCommand { get; }
+        public ICommandEx BackCommand { get; }
+       
         public ObservableRangeCollection<Movie> Movies { get; set; }
-                
+
+        INavigation Navigator { get; }        
+
         public DataServiceConnector service;
         private int page;
         private int listPage;
 
-        public MoviesListViewModel(DataServiceConnector _service, int _listType)
+        private Movie selectedItem;
+        public Movie SelectedItem
+        {
+            get { return selectedItem; }
+            set
+            {
+                SetProperty(ref selectedItem, value);
+                if (selectedItem != null)
+                    ItemSelectedAsync();
+            }
+        }
+
+        public MoviesListViewModel(DataServiceConnector _service, INavigation _navigator, int _listType)
         {
             page = 1;
             service = _service;
             listPage = _listType;
             Movies = new ObservableRangeCollection<Movie>();
-            SelectListTypePage();            
+            Task.Run(async () => { await SelectListTypePage(); });
+            Navigator = _navigator;
+            NextCommand = AsyncCommand.Create("The Next Command", NextPageAsync);
+            BackCommand = AsyncCommand.Create("The Back Command", BackPageAsync);
+            
         }
 
-        public void SelectListTypePage()
+        public async Task NextPageAsync()
+        {
+            MovePage(1);
+            if (Page != 0)
+                await SelectListTypePage();
+        }        
+
+        public async Task BackPageAsync()
+        {
+            MovePage(-1);
+            if (Page != 0)
+                await SelectListTypePage();
+        }
+
+        public void ItemSelectedAsync()
+        {            
+            try
+            {
+                SelectedMovie(service, Navigator, SelectedItem);
+            }
+            catch (Exception ex)
+            {
+                //Insights.Report(ex);
+                //Here we can use a Xamarin insights to report the exception, or one other type of report to catch the exception!
+            }
+        }
+
+        private async Task SelectListTypePage()
         {
             Response<MoviesResult> movies = new Response<MoviesResult>();
             switch(listPage)
             {
                 case 0:
-                    movies = service.GetUpcomingMovies(page);
+                    movies = await service.GetUpcomingMoviesAsync(page);
                     Title = "Upcoming Movies";
                     break;
                 case 1:
-                    movies = service.GetTopRatedMovies(page);
+                    movies = await service.GetTopRatedMoviesAsync(page);
                     Title = "Top Rated Movies";
                     break;
                 default:
@@ -46,17 +95,6 @@ namespace MoviesListProject.ViewModels
             }
             Movies?.Clear();
             Movies.AddRange(movies.Body.Results.ToList<Movie>());
-        }
-
-
-        public void ExecuteLoadNewMoviesCommand(int pageIterator)
-        {
-            if (page + pageIterator != 0)
-            {
-                page += pageIterator;
-                SelectListTypePage();
-            }
-        }
-
+        }        
     }
 }
